Sync deps: libcat
* libcat/libcat@e78448587f696e07db5c102170c03ea50b2a3abf

+++ ext/deps/libcat/include/cat_event.h
@@ -36,7 +36,7 @@ CAT_GLOBALS_STRUCT_BEGIN(cat_event) {
     uv_timer_t deadlock;
     cat_queue_t runtime_shutdown_tasks;
     cat_queue_t defer_tasks;
-    size_t defer_task_count;
+    uint32_t defer_task_count;
 } CAT_GLOBALS_STRUCT_END(cat_event);
 
 extern CAT_API CAT_GLOBALS_DECLARE(cat_event);
